Fix thinblock_tests.cpp
For some reason the deterministic bloom filtering is not entirely
reliable on the release branch and when running on Travis.  We see
an occasional bloom filter false positive when we shouldn't.

There is not big risk here, the bloom filter is only off by one
element but it is a bit of a mystery.  The deterministic behavior
of the bloom filter is governed by seed_insecure_rand().  I have
no idea why seed_insecure_rand() isn't always working as expected
but on other systems works fine, and also we don't see the issue
on the dev branch, at least not yet.