Remove test/fork_tests.cpp
It was implemented to test fork helpers based on median time past. Now
that the helper is based on a specific height they don't make sense
anymore.

Remove also `forkActivateNow` from buip055forc.cpp since the only
place where it was used was in test we are removing with this commit