use cs_objDownloader to lock mapBlocksInFlight (#1022)
* use cs_objDownloader to lock mapBlocksInFlight

* Remove pindex from QueuedBlock

pindex is only ever used when looking up the block height in
GetNodeStateStats(). We can remove it here and rather lookup height by
block hash and since GetNodeStateStats() is executed rarely we
won't have a performance hit.

* Remove tracking of validated headers

This is not a useful metric anymore since we almost all the time,
except under unusual circumstances, will ask for a block without
having the header already. Furthermore, we were not tracking the data
correctly anyway.

More importantly, this would not allow us to finally remove cs_main
from MarkBlocksAsInFlight because of the need to track pindex. So
removing this is yet another step toward finally getting cs_main
out of the request manager.