Fix python and unit tests (#1089)
* Set the minining max block size in excessiveblock_test.cpp

This makes sure we're starting at the right miningsize in case
it was somehow set to a different value in a previous test.

* fix transaction_tests.cpp

assigne the value of 83 to maxdatacarrierbytes in case it was changed
prior to running this set of tests.

* Fix excessive.py so that it works after the may15th fork

* fix mempool_reorg.py to work after fork activation.

* Fix the may152018_forkactivation_2.py

Since the fork time has passed we need to explicity set the pre
fork time when we mined the first set of blocks.

* small fix for mempool_reorg