main.cpp cleanup (#1094)
* Add GUARDED_BY annotations to any maps or data requiring cs_main

There are several items requiring cs_main spread out through the
codebase. This helps devs understand what needs locking and also
gives a compilation warning if rules are not being followed.

* Move BLOCK_DOWNLOAD_WINDOW to the request manager

and make it a public atomic value.

* Make nTimeBestReceived an atomic

* Remove unnecessary cs_blocksequenceid and rely on cs_main

nBlockSequenceId is only used in one place in main.cpp so we
don't need a separate lock for this.

* Add missing cs_main lock to UpdatePreferredDownload

* Move pcoinsdbview to txdb.cpp and txdb.h

* fix format

* Make nPreferrredDownload atomic