Ensure cs_utxo is locked by the caller when a coin reference is returned (#1111)
* Fix locking of AccessCoin() and AccssByTxid()

Both these functions return a reference to a coin in the coins cache
and as a result we have to make sure the caller has locked cs_utxo
for as long as it uses the returned reference.

* Ensure that cs_utxo is locked in CheckTxInputs()

When we call AccessCoin() we must make sure that the returned coin
reference is still locked for as long as we need it.

* Temporarily lock cs_utxo while we copy the amount and scriptPubKey

This way we minimize the time cs_utxo is locked and prevent it from
being locked while the scripts and sigs are being checked.

* Add clarifying comment as to why we do the copy

We need the copy becasue we can't depend on "coin" still being
valid.

We could do the copy further up and then release the lock before
taking cs_main however that would require a copy each and every
time we CheckTxInpts() when in reality this section of the code
gets executed rarely.