Remove unnecessary vNodes locks (#1134)
* Move requester.SendRequests() up before we release the node refs.

This way we don't have to take so many vNodes locks during
SendRequests().

* Remove unnecessary vNodes locks

* Add additional comments explaining the why we don't need to lock vNodes
when releasing refs and also, at times, when adding refs in the
requestManager.

and also tidy up the formatting in ThreadMessageHandler()

* This comment is actually not true

We don't have to keep the refs here all the way until the end
because the requestManager has its own set of refs as well, however,
it seems sensisble to do so anyway as a last step.

* Add DbgAssert() to ensure that we have a ref to the node already

and therefore don't need to lock vNodes.

* DbgAssert for nRefCount

nRefCount should always be greater than zero before we release a ref.