Use block height rather than time to compare against BIP16Height (#1170)
* Use block height rather than time to compare against BIP16Height

In the process of porting a bunch PR from core in the context of the IBD
code refactoring we erroneously use `GetBlockTime` rather than height to
compare against `consensusparams.BIP16Height`.

* Move p2sh-invalid-txn after heigth ~20K

The p2sh rules start to be enforced starting from consensus.BIP16Height,
i.e. 173805. Hence  to check if this is the case we need to try adding an
invalid p2sh txn to the mempool only after such new rules are enforced.