Randomize the graphene and thinblock timers.
By randomizing the timeouts either plus or minus 2 seconds
the propagation of graphene and thinblocks is enhanced near
the margins of the BU network, where nodes can receive the block
announcment at the same time. When such nodes are connected to
each other they previously would timeout and default to either an
xthin, or full block download.  Also, for new users, or even for
testing purposes it is also helpful in seeing a greater proportion
of graphene blocks being send or received when there a limited
number of graphene peers available.