Remove old xthin code which is no longer used (#1214)
* Remove connect-thinblock and connect-thinblock-force functionality.

This functionality is a fair amount of technical debt which came as
a result of the first first iterations of XThins and was originally
used to connect xthin nodes to each other and also used in the GFC testing.

We now do not need any way to connect do each other because of the
thinblock and graphene timers, but we can used  -addnode which is a much
better way to connect nodes together.
We have really no need for the connect-thinblock-force functionality we
used in GFC testing, and there is a fair amount of code clutter here we
can remove which  results in a much cleaner RequestBlock() in the RequestManager.

If we ever need to do and testing in the future to force xthins for graphene
blocks from special peers it will be easier and to just jerry rig something
for testing rather than keeping all this code around.

* fix compile warning when comparing int64_t and uint64_t