Display advanced block propagation technology (XThin/Graphene) statistics in the Debug Window UI (#1221)
* Separately track inbound and outbound XThin blocks

This is so we can maintain and display the total send/receive count, in addition to the last 24-hour statistics (matching graphene).

* Correct thinblock qa test

Test relies on exact output string format of 'thinblockstats'.'summary' field
which was changed in the previous commit

* Add quick stats struct to ThinBlock and helper methods to fill the data

This is to allow grabbing multiple statistics in one go for display in the UI while only taking a single lock on cs_thinblockstats
The helper methods compute in exactly the same way the existing string methods do, and are intended to replace the internal computations to properly follow DRY methodology (subsequent commit).
NOTE: Set up to use clang static analysis LOCK macros.

* Update existing ThinBlock string methods to use new internal methods instead

This is such that any computations use a single algorithm

* Update statistics map iterators to C++11 range based for loops

* Update all 24-hour average statistics to expirestats() on all maps used in computation

Some computations called expire stats on only one of the two maps which could potentially lead to the second map containing more data entries than first map, skewing the results.
Other computations did not expire stats at all on the map used in the computation which could potentially lead to the computed stat spanning slightly more than the expected 24-hour period.

* Hook up stat update to the UI

* Add labels to UI for block propagation stats

* Add quick stats struct to Graphene and helper methods to fill the data

This is to allow grabbing multiple statistics in one go for display in the UI while only taking a single lock on cs_graphenestats
The helper methods compute in exactly the same way the existing string methods do, and are intended to replace the internal computations to properly follow DRY methodology (subsequent commit).
NOTE: Set up to use clang static analysis LOCK macros.

* Update existing Graphene string methods to use new internal methods instead

This is such that any computations use a single algorithm