Add LOCK(cs_utxo) to the remaining CCoinsViewDB methods
It appears that leveldb has issues with doing db.Read(). A core
was produced in which it appears that a block had just finished
flushing to the dbcache, however a late incoming txn was then
trying to access the coins in the dbcache which went missing during
it's attempt to db.Read() from leveldb.  Levedb doesn't seem to
be able to handle this, although leveldb is supposed to be ACID.

It looks, from the backtrace, that the key is being looked up in
the leveldb memtable and it is found, however, when trying to then
acess the memory it segfaults.