Remove trickle logic (#1299)
* Remove trickle logic

* Fix send INV logic to make sure all INV's are being sent.

There was a bug which was truncating a good portion of INV's when
we had to send more than 1000.

* Tidy up bool naming convention

* Make sure to send all INV's before returning.

It is possible, that under a very heaving load of incoming transactions
that we could fall behind in sending out INVs if we limited each pass
through by only 1000 and vInventoryToSend could grow quite large. Therefore
make sure to send all INV's in our queue before returning.

* Make sure to hold cs_inventory when we erase from vInventoryToSend

* Move MAX_INV_TO_SEND into main.h and make it a static