UI - Adjust word wrap for peer details (#1303)
Turn off for header and put node ID on second line so that header always has full width to display the IP (handles long IPv6 addresses).
Turn on for user agent and services strings.
NOTE: Required adding in a space to the separators between services as word wrap only works on whitespace characters for QLabels